- renamed User Rights - added Get Userrights - made it possible to create admin users - changed  changeRole to settings view and corrected the validation - fixed error with Microcontroller Serializer/added new one

+++ backend/app/e_paul_smarthome_system/views/userView.py
@@ -1,5 +1,8 @@
 from ..model.user import User
 from ..model.account import Account
+from ..model.group import Group
+from ..model.port import Port
+from ..model.groupPort import GroupPort
 from ..serializer.userSerializer import UserSerializer, UserDetailSerializer, UserEditSerializer
 
 from rest_framework.views import APIView
@@ -17,12 +20,23 @@ class SingleUserView(APIView):
     queryset = User.objects.all()
 
     def get(self, request, userId):
-        user = User.objects.get(pk = userId)
+        try:
+            user = User.objects.get(pk = userId)
+        except User.DoesNotExist:
+            return Response(status = 400)
+        
         serializer = UserDetailSerializer(user)
         return Response(serializer.data, status = 200)
         
     def delete(self, request, userId):
-        user = User.objects.get(pk = userId)
+        try:
+            user = User.objects.get(pk = userId)
+        except User.DoesNotExist:
+            return Response(status = 400)
+        
+        if user.role == "superuser":
+            return Response('Superuser kann nicht gelöscht werden', status=400)
+        
         user.delete()
         return Response('User erfolgreich gelöscht')
